% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{as.string.c}
\alias{as.string.c}
\title{as.string.c}
\usage{
as.string.c(vect, quote = TRUE)
}
\arguments{
\item{vect}{a vector}

\item{quote}{a logical value. If TRUE, the values on the vector will be surrounded by quotes.}
}
\value{
a text
}
\description{
Creates a string representative of a vector
}
\examples{
as.string.c(c("A", "B", "C"))
as.string.c(c(5, 6, 7))
as.string.c(c(5, 6, 7), quote = FALSE)
as.string.c(iris$Species)

}
