% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{as_string_c}
\alias{as_string_c}
\title{as_string_c}
\usage{
as_string_c(vect, quote = TRUE)
}
\arguments{
\item{vect}{a vector with values}

\item{quote}{a logical value. If TRUE, the values on the vector will be surrounded by quotes.}
}
\description{
creates a string representative of a vector
}
\examples{
as_string_c(c("A", "B", "C"))
as_string_c(c(5, 6, 7))
as_string_c(c(5, 6, 7), quote = FALSE)
as_string_c(iris$Species)

}
