% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boosting_utils.R
\name{boosting_prediction}
\alias{boosting_prediction}
\title{boosting_prediction}
\usage{
boosting_prediction(model, test.data, n.trees = 50)
}
\arguments{
\item{model}{boosting model(gbm).}

\item{test.data}{dataframe.}

\item{n.trees}{number of trees used in the prediction.}
}
\description{
generates the prediction of a boosting model.
}
\seealso{
\code{\link[gbm]{gbm}}
}
