% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boosting_utils.R
\name{calibrate_boosting}
\alias{calibrate_boosting}
\title{calibrate_boosting}
\usage{
calibrate_boosting(data)
}
\arguments{
\item{data}{the name of the learning data.}
}
\description{
helps to get the maximum of n.minobsinnode and bag.fraction values with which no error is generated in the model.
}
\examples{
calibrate_boosting(iris)

}
\seealso{
\code{\link[gbm]{gbm}}
}
