% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_manipulation.R
\name{categorical_distribution}
\alias{categorical_distribution}
\title{categorical_distribution}
\usage{
categorical_distribution(var)
}
\arguments{
\item{var}{a vector with the data for the categorical distribution chart.}
}
\description{
makes the graph of the categorical distribution.
}
\examples{
categorical_distribution(iris$Species)

}
