% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tools.R
\name{categorical_summary}
\alias{categorical_summary}
\title{categorical_summary}
\usage{
categorical_summary(data, variable)
}
\arguments{
\item{data}{a data.frame with the data for analysis.}

\item{variable}{the name of the variable for analysis.}
}
\description{
generates the fields for individual categorical analysis.
}
\examples{
if(interactive()) {
  library(shiny)
  library(DT)
  shinyApp(ui = fluidPage(fluidRow(uiOutput("resumen"))),
           server = function(input, output) {
                         output$resumen = renderUI(categorical_summary(iris, "Species"))
           })
}
}
