% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_manipulation.R
\name{chunk}
\alias{chunk}
\title{chunk}
\usage{
chunk(content = "")
}
\arguments{
\item{content}{the content to be inserted.}
}
\description{
locks the data into a rmarkdown chunk.
}
\keyword{internal}
