% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_manipulation.R
\name{clean_report}
\alias{clean_report}
\title{clean_report}
\usage{
clean_report()
}
\description{
clean the full report.
}
\examples{
new_report(iris, 'iris')
get_report()
clean_report()
get_report()

}
