% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{code_load}
\alias{code_load}
\title{code_load}
\usage{
code_load(
  row.names = TRUE,
  path = NULL,
  sep = ";",
  sep.dec = ",",
  header = TRUE,
  d.o = "datos.originales",
  d = "datos"
)
}
\arguments{
\item{row.names}{a logical value indicating whether the data has row names.}

\item{path}{the path of the file.}

\item{sep}{the column separator in the file.}

\item{sep.dec}{the decimal separator in the file.}

\item{header}{a logical value indicating whether the file has a header.}

\item{d.o}{the name of the original data.}

\item{d}{the name of the current data.}
}
\description{
generates data reading code.
}
\examples{
code_load(TRUE, "MY/PATH/FILE.csv")

}
