% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{code_summary}
\alias{code_summary}
\title{code_summary}
\usage{
code_summary(data = "datos")
}
\arguments{
\item{data}{the name of the current data.}
}
\description{
creates the code for the basic summary of variables.
}
\examples{
x <- code_summary('iris')
exe(x)

}
