% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{code_transf}
\alias{code_transf}
\title{code_transf}
\usage{
code_transf(variable, new.type, d.o = "datos.originales", d = "datos")
}
\arguments{
\item{variable}{the name of the variable to be converted.}

\item{new.type}{the new type of the variable. Can be categorical, numerical or disjunctive. ('categorico', 'numerico', 'disyuntivo')}

\item{d.o}{the name of the original data.}

\item{d}{the name of the current data.}
}
\description{
generate code to transform data.
}
\examples{
iris2 <- iris
x <-code_transf('Species', 'disyuntivo', 'iris', 'iris2')
exe(x)
head(iris2)

}
