% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{colnames.empty}
\alias{colnames.empty}
\title{colnames.empty}
\usage{
colnames.empty(data)
}
\arguments{
\item{data}{the dataset}
}
\value{
a vector with the names of the columns or an empty string if the data is NULL
}
\description{
colnames.empty
}
\examples{
colnames.empty(iris)
colnames.empty(NULL)

}
