% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{colnames_empty}
\alias{colnames_empty}
\title{colnames_empty}
\usage{
colnames_empty(data)
}
\arguments{
\item{data}{the dataset.}
}
\value{
a vector with the names of the columns or an empty string if the data is NULL.
}
\description{
colnames_empty
}
\examples{
colnames_empty(iris)
colnames_empty(NULL)

}
