% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{combine.names}
\alias{combine.names}
\title{combine.names}
\usage{
combine.names(x, y, sep = ".")
}
\arguments{
\item{x}{a vector to combine with y. The combination is grouped by this parameter.}

\item{y}{a vector to combine with x.}

\item{sep}{a string with the separator characters.}
}
\value{
a vector with the combination of x and y.
}
\description{
combine.names
}
\examples{
x = c("A", "B", "C")
y = c("1", "2", "3")
combine.names(x, y)

}
