% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{comparative_table}
\alias{comparative_table}
\title{comparative_table}
\usage{
comparative_table(sel, indices)
}
\arguments{
\item{sel}{the selection of the models to be shown.}

\item{indices}{the values to be shown.}
}
\description{
creates the comparison table.
}
\examples{
models <- list('knnl-mode1' = list(0.11,0.22,0.33,0.44),
               'dtl-mode2'  = list(0.12,0.23,0.34,0.45),
               'rfl-mode1'  = list(0.51,0.42,0.13,0.24))
sel <- c("K Vecinos Más Cercanos-mode1", "Bosques Aleatorios-mode1")
comparative_table(sel, models)

}
