% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{cor_model}
\alias{cor_model}
\title{cor_model}
\usage{
cor_model(data = "datos")
}
\arguments{
\item{data}{the name of the current data.}
}
\description{
generates the code to calculate the correlation matrix.
}
\examples{
x <- cor_model('iris')
exe(x)
correlacion

}
