% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{correlations_plot}
\alias{correlations_plot}
\title{correlations_plot}
\usage{
correlations_plot(method = "circle", type = "lower")
}
\arguments{
\item{method}{the visualization method of correlation matrix to be used.}

\item{type}{display full matrix, lower triangular or upper triangular matrix.}
}
\description{
generates the code of the correlation chart.
}
\examples{
x <- cor_model('iris')
exe(x)
print(correlacion)

x <- correlations_plot()
exe(x)

}
\seealso{
\code{\link[corrplot]{corrplot}}
}
