% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{def_code_cat}
\alias{def_code_cat}
\title{def_code_cat}
\usage{
def_code_cat(data = "datos", variable, language = "es")
}
\arguments{
\item{data}{the name of the current data.}

\item{variable}{the name of the variable for the categorical distribution chart.}

\item{language}{the language to choose. It can be "es" or "en".}
}
\description{
def_code_cat
}
\examples{
def_code_cat('iris', 'Species')

}
