% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{def_code_num}
\alias{def_code_num}
\title{def_code_num}
\usage{
def_code_num(data = "datos", variable, color = "red")
}
\arguments{
\item{data}{the name of the current data.}

\item{variable}{the name of the variable for the numerical distribution chart.}

\item{color}{the color of the chart.}
}
\description{
def_code_num
}
\examples{
x <- def_code_num('iris', 'Petal.Length')
exe(x)

}
