% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{default_disp}
\alias{default_disp}
\title{default_disp}
\usage{
default_disp(data = "datos", vars = NULL, color = "#FF0000AA")
}
\arguments{
\item{data}{the name of the current data.}

\item{vars}{a vector of length 2 or 3 with the names of the variables for the graph.}

\item{color}{the color of the dots on the chart.}
}
\description{
default_disp
}
\examples{
library(scatterplot3d)

x <- default_disp('iris', c('Sepal.Length', 'Sepal.Width'))
exe(x)

x <- default_disp('iris', c('Sepal.Length', 'Sepal.Width', 'Petal.Length'))
exe(x)

}
