% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{disjunctive.data}
\alias{disjunctive.data}
\title{disjunctive.data}
\usage{
disjunctive.data(data, vars)
}
\arguments{
\item{data}{the dataset to be converted}

\item{vars}{a vector with the name of columns}
}
\value{
a dataset
}
\description{
Convert the columns selected to disjunctive
}
\examples{
disjunctive.data(iris, "Species")
}
