% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{disjunctive_data}
\alias{disjunctive_data}
\title{disjunctive_data}
\usage{
disjunctive_data(data, vars)
}
\arguments{
\item{data}{the dataset to be converted.}

\item{vars}{a vector with the name of columns.}
}
\description{
convert the columns selected to disjunctive.
}
\examples{
disjunctive_data(iris, "Species")

}
