% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnTrivial.R
\name{.regressoR.learnTrivial}
\alias{.regressoR.learnTrivial}
\title{Check if a Regression Problem has a Trivial Solution}
\usage{
.regressoR.learnTrivial(x, y)
}
\arguments{
\item{x}{the \code{x} coordinates, i.e., the input values}

\item{y}{the \code{y} coordinates, i.e., the output values}
}
\value{
an instance of \code{\link{FittedModel}} which represents the
  relationship between the \code{x} and \code{y} values
}
\description{
Find a trivial solution for a regression problem, if any.
}
