% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_regression_trees_utils.R
\name{dt_model}
\alias{dt_model}
\title{dt_model}
\usage{
dt_model(data, variable.pred, minsplit = 20, maxdepth = 15)
}
\arguments{
\item{data}{dataframe}

\item{variable.pred}{the name of the variable to be predicted.}

\item{minsplit}{the minsplit parameter of the model.}

\item{maxdepth}{the maxdepth parameter of the model.}
}
\description{
generates a decision trees model.
}
\seealso{
\code{\link[rpart]{rpart}}
}
