% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{dt_plot}
\alias{dt_plot}
\title{dt_plot}
\usage{
dt_plot(model.var = "modelo.dt")
}
\arguments{
\item{model.var}{the name of the variable that stores the resulting prediction.}
}
\description{
makes the graph of the tree.
}
\examples{
\dontrun{
library(rpart)

x <- dt_model('iris', 'Petal.Length', model.var = 'model_dt')
exe(x)
print(model_dt)

x <- dt_plot('model_dt')
exe(x)
}
}
