% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_regression_trees_utils.R
\name{dt_prediction}
\alias{dt_prediction}
\title{dt_prediction}
\usage{
dt_prediction(model, test.data)
}
\arguments{
\item{model}{a decision trees model(rpart).}

\item{test.data}{dataframe.}
}
\description{
generates the prediction of the decision trees model.
}
