% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_rdim_utils.R
\name{e.rdim.rmse}
\alias{e.rdim.rmse}
\title{e.rdim.rmse}
\usage{
e.rdim.rmse(modelo, ncomp, titles = c("RMSE", "Componente"))
}
\arguments{
\item{modelo}{a dimension reduction model.}

\item{ncomp}{the optimum number of components.}

\item{titles}{labels on the chart}
}
\value{
echarts4r plot
}
\description{
graph the root mean square error of cross validation according to components used.
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
