% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_rdim_utils.R
\name{e.rdim.vare}
\alias{e.rdim.vare}
\title{plot_pred_rd}
\usage{
e.rdim.vare(modelo, ncomp, titles = c("Varianza Explicada", "Componente"))
}
\arguments{
\item{modelo}{a dimension reduction model.}

\item{ncomp}{the optimum number of components.}

\item{titles}{labels on the chart}
}
\value{
echarts4r plot
}
\description{
graph of variance explained in the predictors according to components used.
}
\author{
Ariel Arroyo <luis.ariel.arroyo@promidat.com>
}
