% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_boost_utils.R
\name{e_boost_evol_error}
\alias{e_boost_evol_error}
\title{Error Evolution}
\usage{
e_boost_evol_error(modelo, label = "Iterations")
}
\arguments{
\item{modelo}{a adabag model.}

\item{label}{a label plot.}
}
\value{
echarts4r plot
}
\description{
Error Evolution
}
\examples{
model <- traineR::train.gbm(Sepal.Length~., data = iris, 
  distribution = "gaussian", n.trees = 5, shrinkage = 0.01)
e_boost_evol_error(model, iris)

}
\author{
Joseline Quiros <joseline.quiros@promidat.com>
}
