% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_boost_utils.R
\name{e_boost_importance}
\alias{e_boost_importance}
\title{Var importance Adabag}
\usage{
e_boost_importance(modelo)
}
\arguments{
\item{modelo}{a adabag model.}
}
\value{
echarts4r plot
}
\description{
Var importance Adabag
}
\examples{
model <- traineR::train.gbm(Sepal.Length~., data = iris, 
  distribution = "gaussian", n.trees = 5, shrinkage = 0.01)
e_boost_importance(model)

}
\author{
Joseline Quiros <joseline.quiros@promidat.com>
}
