% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_penalized_Regression_utils.R
\name{e_coeff_landa}
\alias{e_coeff_landa}
\title{e_coeff_landa}
\usage{
e_coeff_landa(
  cv.glm,
  log.lambda = NULL,
  titles = c("Coeficientes", "Seleccionado", "Automatico")
)
}
\arguments{
\item{cv.glm}{a cv.glmnet model.}

\item{log.lambda}{number that specifies the logarithm of the selected lambda}

\item{titles}{labels on the chart}
}
\value{
echarts4r plot
}
\description{
Graph the coefficients and lambdas of a cv.glmnet model
}
\seealso{
\code{\link[glmnet]{cv.glmnet}}
}
\author{
Ariel Arroyo <luis.ariel.arroyo@promidat.com>
}
