% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_penalized_Regression_utils.R
\name{e_posib_lambda}
\alias{e_posib_lambda}
\title{e_posib_lambda}
\usage{
e_posib_lambda(
  cv.glm,
  log.lambda = NULL,
  titles = c("Error Cuadratico Medio", "Curva Inferior", "Curva Superior",
    "Seleccionado", "Automatico", "Coeficientes Distintos de Cero")
)
}
\arguments{
\item{cv.glm}{a cv.glmnet model.}

\item{log.lambda}{number that specifies the logarithm of the selected lambda}

\item{titles}{labels on the chart}
}
\value{
echarts4r plot
}
\description{
Graph a cv.glmnet model
}
\seealso{
\code{\link[glmnet]{cv.glmnet}}
}
\author{
Ariel Arroyo <luis.ariel.arroyo@promidat.com>
}
