% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_manipulation.R
\name{error.variables}
\alias{error.variables}
\title{error.variables}
\usage{
error.variables(num = T)
}
\arguments{
\item{num}{if TRUE shows a message of missing numerical variables, if FALSE shows a message of missing categorical variables.}
}
\description{
Draws an error of missing data
}
\examples{
error.variables(TRUE)
error.variables(FALSE)
}
