% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_manipulation.R
\name{error_plot}
\alias{error_plot}
\title{error_plot}
\usage{
error_plot(msg)
}
\arguments{
\item{msg}{the message to be displayed in the graph}
}
\description{
makes a warning graphic
}
\examples{
error_plot("My Warning")

}
