% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_manipulation.R
\name{error_variables}
\alias{error_variables}
\title{error_variables}
\usage{
error_variables(num = T)
}
\arguments{
\item{num}{if TRUE shows a message of missing numerical variables, if FALSE shows a message of missing categorical variables.}
}
\description{
draws an error of missing data.
}
\examples{
error_variables(TRUE)
error_variables(FALSE)

}
