% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{exe}
\alias{exe}
\title{exe}
\usage{
exe(...)
}
\arguments{
\item{...}{one or more texts to be concatenated and executed}
}
\value{
the result of the execute
}
\description{
Concat and execute a text in R
}
\examples{
exe("5+5")
exe("5","+","5")
exe("plot(iris$Species)")

}
