% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{extract.code}
\alias{extract.code}
\title{extract.code}
\usage{
extract.code(funcion)
}
\arguments{
\item{funcion}{the name of the function to be extracted}
}
\value{
the code in text form
}
\description{
Gets the code of a function in text form
}
\examples{
extract.code("cat")
extract.code("plot")

parse(text = extract.code("plot"))

}
