% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{fisher_calc}
\alias{fisher_calc}
\title{fisher_calc}
\usage{
fisher_calc(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector with the data to make the calculation.}

\item{na.rm}{a logical value indicating whether the NAs have to be eliminated.}
}
\description{
calculate the fisher skewness.
}
\examples{
fisher_calc(iris$Petal.Length)

}
