% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{general.indices}
\alias{general.indices}
\title{general.indices}
\usage{
general.indices(real, prediccion)
}
\arguments{
\item{real}{the real values in traning-testing.}

\item{prediccion}{the prediction values in traning-testing.}
}
\value{
a list with Correlation, Relative Error, Mean Absolute Error, Root Mean Square Error
}
\description{
calculates indices to measure accuracy of a model
}
\examples{
real <- rnorm(45)
prediction <- rnorm(45)
model <- "KNN"
general.indices(real, prediction)

}
