% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_manipulation.R
\name{get_env_report}
\alias{get_env_report}
\title{get_env_report}
\usage{
get_env_report()
}
\description{
gets the environment where the list is found with the report.
}
\examples{
e <- get_env_report()
e$codigo.reporte

}
