% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_manipulation.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{gg_color_hue}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{an integer specifying the number of colors to create.}
}
\value{
color-coded vector
}
\description{
create colors.
}
\examples{
col <- gg_color_hue(3)
plot(iris$Species, col = col)

}
