% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_manipulation.R
\name{importance_plot_rf}
\alias{importance_plot_rf}
\title{importance_plot_rf}
\usage{
importance_plot_rf(
  model.rf,
  titles = c("Importancia de Variables Según el Porcentaje de Incremento del MSE",
    "Aumento porcentual del error cuadrático medio", "Variable")
)
}
\arguments{
\item{model.rf}{a random forest model.}

\item{titles}{labels on the chart}
}
\value{
echarts4r plot
}
\description{
graphs the importance of variables for the random forest model according to the percentage increase in mean square error.
}
\examples{
library(randomForest)
x <- rf_model('iris', 'Petal.Length')
exe(x)
importance_plot_rf(modelo.rf)

}
\seealso{
\code{\link[randomForest]{randomForest}}
}
\author{
Ariel Arroyo <luis.ariel.arroyo@promidat.com>
}
