% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tools.R
\name{infoBoxPROMiDAT}
\alias{infoBoxPROMiDAT}
\title{infoBoxPROMiDAT}
\usage{
infoBoxPROMiDAT(title, value, icon)
}
\arguments{
\item{title}{title of the box}

\item{value}{the content of the box}

\item{icon}{icon of the box}
}
\value{
shiny.tag object
}
\description{
create a box for the information tab
}
\details{
this function only work correctly on the server side because they need the css and js file.
}
\keyword{internal}
