% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_regresor.R
\name{init_regresor}
\alias{init_regresor}
\title{This function will start regresoR}
\usage{
init_regresor()
}
\value{
Nothing
}
\description{
An interactive 'Shiny' application for data prediction.
}
\details{
Start regresoR

This starts the regresoR application on the user's local computer.
}
\examples{
 if(interactive()){
   init_regresor()
 }
}
\keyword{predictoR}
