% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tools.R
\name{inputRadio}
\alias{inputRadio}
\title{inputRadio}
\usage{
inputRadio(inputId, value, isSelected)
}
\arguments{
\item{inputId}{the id of the radio button}

\item{value}{a text value of the radio button}

\item{isSelected}{logical. If TRUE the radio button is selected}
}
\value{
shiny.tag object
}
\description{
create a radio button input
}
\details{
this function only work correctly on the server side because they need the css and js file.
}
\keyword{internal}
