% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_manipulation.R
\name{insert_report}
\alias{insert_report}
\title{insert_report}
\usage{
insert_report(id, title = NA, ..., interpretation = TRUE,
  is.chunk = TRUE, add = FALSE)
}
\arguments{
\item{id}{a string with the key of what is inserted in the report.}

\item{title}{the title of the content, if there is no title is NA.}

\item{...}{the content to be inserted.}

\item{interpretation}{a logical value indicating whether a label has to be inserted for interpretation.}

\item{is.chunk}{a logical value indicating whether the content has to be enclosed in a chunk.}

\item{add}{a logical value indicating if the content has to be added to what is a before.}
}
\description{
inserts an element in the report in the current section.
}
\examples{
new_report(iris, "iris")
insert_report("1_part", 'Title 1', 'head(iris)\\n', 'summary(iris)')
get_report()
clean_report()

}
