% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{kkn_model}
\alias{kkn_model}
\title{kkn_model}
\usage{
kkn_model(
  data = "datos.aprendizaje",
  variable.pred = NULL,
  scale = TRUE,
  kmax = 7,
  kernel = "optimal",
  model.var = "modelo.knn",
  distance = 2
)
}
\arguments{
\item{data}{the name of the learning data.}

\item{variable.pred}{the name of the variable to be predicted.}

\item{scale}{the scale parameter of the model.}

\item{kmax}{the kmax parameter of the model.}

\item{kernel}{the kernel parameter of the model.}

\item{model.var}{the name of the variable that stores the resulting model.}

\item{distance}{the distance parameter of the model.}
}
\description{
generates the code to create the k nearest neighbors model.
}
\examples{
library(kknn)
x <- kkn_model('iris', 'Petal.Length')
exe(x)
print(modelo.knn)

}
\seealso{
\code{\link[kknn]{train.kknn}}
}
