% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tools.R
\name{labelInput}
\alias{labelInput}
\title{labelInput}
\usage{
labelInput(inputId, value = "")
}
\arguments{
\item{inputId}{The id of the label.}

\item{value}{Default Value of the label (optional). Default value is "".}
}
\value{
shiny.tag object
}
\description{
label with identifier for language change
}
\details{
this function only work correctly on the server side because they need the css and js file.
}
\keyword{internal}
