% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_carga_datos.R
\name{mod_carga_datos_server}
\alias{mod_carga_datos_server}
\title{carga_datos Server Function}
\usage{
mod_carga_datos_server(input, output, session, updateData, modelos)
}
\description{
carga_datos Server Function
}
\keyword{internal}
