% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ind_nuevos.R
\name{mod_ind_nuevos_server}
\alias{mod_ind_nuevos_server}
\title{ind_nuevos Server Function}
\usage{
mod_ind_nuevos_server(input, output, session, newCases, updateData2, codedioma)
}
\description{
ind_nuevos Server Function
}
\keyword{internal}
