% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_new_data_predictions.R
\name{mod_new_data_predictions_server}
\alias{mod_new_data_predictions_server}
\title{new_data_predictions Server Function}
\usage{
mod_new_data_predictions_server(input, output, session, updateData, new.data)
}
\description{
new_data_predictions Server Function
}
\keyword{internal}
