% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_manipulation.R
\name{new_section_report}
\alias{new_section_report}
\title{new_section_report}
\usage{
new_section_report()
}
\description{
creates a new section in the report, this way you can overwrite keys and delete an element only affects the current section.
}
\examples{
new_report(iris, 'iris')
insert_report('1_part', 'Title 1', 'head(iris)\\n', 'summary(iris)')
get_report()

remove_report_elem('1_part')
get_report()

new_section_report()
insert_report('1_part', 'Title 1', 'head(iris)\\n', 'summary(iris)')
get_report()

new_section_report()
insert_report('1_part', 'Title 1', 'head(iris)\\n', 'summary(iris)')
get_report()

remove_report_elem('1_part')
get_report()

clean_report()

}
