% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_neural_networks_utils.R
\name{nn_model}
\alias{nn_model}
\title{nn_model}
\usage{
nn_model(data, variable.pred, hidden = c(1), threshold = 0.1, stepmax = 2000)
}
\arguments{
\item{data}{dataframe}

\item{variable.pred}{the name of the variable to be predicted.}

\item{hidden}{a vector of integers specifying the number of hidden neurons (vertices) in each layer.}

\item{threshold}{a numeric value specifying the threshold for the partial derivatives of the error function as stopping criteria.}

\item{stepmax}{the maximum steps for the training of the neural network. Reaching this maximum leads to a stop of the neural network's training process.}
}
\description{
generates the code to create the neural network model.
}
\seealso{
\code{\link[neuralnet]{neuralnet}}
}
