% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{nn_plot}
\alias{nn_plot}
\title{nn_plot}
\usage{
nn_plot(model.var = "modelo.nn")
}
\arguments{
\item{model.var}{the name of the variable that stores the resulting model.}
}
\description{
generates the code to create the graph of the neural network.
}
\examples{
\dontrun{
library(neuralnet)
library(dummies)
library(dplyr)

x <- nn_model('iris', 'Petal.Length','modelo.nn', 'mean.nn', 'sd.nn', 0.05, 2000, 3, 10, 10, 10)
exe(x)

x <- nn_plot('modelo.nn')
exe(x)
}
}
