% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{nn_prediction}
\alias{nn_prediction}
\title{nn_prediction}
\usage{
nn_prediction(
  data = "datos.prueba",
  variable.pred = NULL,
  model.var = "modelo.nn",
  pred.var = "prediccion.nn",
  mean.var = "mean.nn",
  sd.var = "sd.nn"
)
}
\arguments{
\item{data}{the name of the test data.}

\item{variable.pred}{the name of the variable to be predicted.}

\item{model.var}{the name of the variable that stores the resulting model.}

\item{pred.var}{the name of the variable that stores the resulting prediction.}

\item{mean.var}{the name of the variable that stores the mean of the columns.}

\item{sd.var}{the name of the variable that stores the standard deviation of the columns.}
}
\description{
generates the code to create the prediction of the neural network model.
}
\examples{
\dontrun{
library(neuralnet)
library(dummies)
library(dplyr)

x <- nn_model('iris', 'Petal.Length','modelo.nn', 'mean.nn', 'sd.nn', 0.05, 2000, 3, 30, 50, 30)
exe(x)

x <- nn_prediction('iris', 'Petal.Length')
exe(x)
print(prediccion.nn)
}
}
\seealso{
\code{\link[neuralnet]{compute}}
}
