% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{normal_default}
\alias{normal_default}
\title{normal_default}
\usage{
normal_default(
  data = "datos",
  vars = NULL,
  color = "#00FF22AA",
  labelcurva = "Curva Normal"
)
}
\arguments{
\item{data}{the name of the current data.}

\item{vars}{the variable for analysis. It has to be numeric.}

\item{color}{the color of the histogram.}

\item{labelcurva}{label for the curve.}
}
\description{
generates the code of the normality test.
}
\examples{
x <- normal_default('iris', 'Sepal.Length')
exe(x)

}
