% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_manipulation.R
\name{numerical_distribution}
\alias{numerical_distribution}
\title{numerical_distribution}
\usage{
numerical_distribution(var, var.name, color)
}
\arguments{
\item{var}{a vector with the data for the numerical distribution chart.}

\item{var.name}{the name of the variable.}

\item{color}{the color of the chart.}
}
\description{
makes the graph of the numerical distribution.
}
\examples{
numerical_distribution(iris[,'Sepal.Length'], 'Sepal.Length', 'red')

}
