% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tools.R
\name{numerical_summary}
\alias{numerical_summary}
\title{numerical_summary}
\usage{
numerical_summary(data, variable)
}
\arguments{
\item{data}{a data.frame with the data for analysis.}

\item{variable}{the name of the variable for analysis.}
}
\description{
generates the fields for individual numerical analysis.
}
\examples{
if(interactive()) {
  library(shiny)
  library(DT)
  shinyApp(ui = fluidPage(fluidRow(uiOutput("resumen"))),
           server = function(input, output) {
                         output$resumen = renderUI(numerical_summary(iris, "Sepal.Width"))
           })
}
}
