% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{options_regressor}
\alias{options_regressor}
\title{options_regressor}
\usage{
options_regressor(...)
}
\arguments{
\item{...}{any options can be defined, using name = value or a character string holding an option name.}
}
\description{
options_regressor
}
\examples{
options_regressor("language")
options_regressor(language = "en")
options_regressor("language")

}
