% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_manipulation.R
\name{order_report}
\alias{order_report}
\title{order_report}
\usage{
order_report(list_report)
}
\arguments{
\item{list_report}{la lista del con los elementos del reporte.}
}
\description{
sort the report list according to the 'regressoR' logic.
}
\keyword{internal}
