% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Predictive_Power_utils.R
\name{pairs_power}
\alias{pairs_power}
\title{pairs_power}
\usage{
pairs_power(data)
}
\arguments{
\item{data}{A DataFrame}
}
\description{
Generate a pair chart
}
\seealso{
\code{\link[psych]{pairs.panels}}
}
