% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{partition_code}
\alias{partition_code}
\title{partition_code}
\usage{
partition_code(data = "datos", p = 50, variable = NULL,
  semilla = 5, perm.semilla = FALSE)
}
\arguments{
\item{data}{the name of the current data.}

\item{p}{the percentage of data for the learning data.}

\item{variable}{the name of the variable to be predicted.}

\item{semilla}{a number with the random seed.}

\item{perm.semilla}{a logical value indicating whether the random seed should be used.}
}
\description{
creates the partition code for testing and learning data.
}
\examples{
x <- partition_code('iris', 75, 'Species', 555, TRUE)
exe(x)
head(datos.aprendizaje)
head(datos.prueba)

}
