% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{plot_coef_lambda}
\alias{plot_coef_lambda}
\title{plot_coef_lambda}
\usage{
plot_coef_lambda(model.var = "modelo.rlr", lambda = NULL, cv.var = "cv.glm")
}
\arguments{
\item{model.var}{the name of the variable that stores the resulting model.}

\item{lambda}{a numerical value in case you don't want to use the optimal lambda.}

\item{cv.var}{the variable that stores the optimal lambda.}
}
\description{
generates the code to plot the penalized regression coefficients.
}
\examples{
library(glmnet)
x <- rlr_model('iris', 'Petal.Length')
exe(x)

x <- plot_coef_lambda('modelo.rlr')
exe(x)

}
