% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_manipulation.R
\name{plot_pred_rd}
\alias{plot_pred_rd}
\title{plot_pred_rd}
\usage{
plot_pred_rd(model)
}
\arguments{
\item{model}{a dimension reduction model.}
}
\description{
graph of variance explained in the predictors according to components used.
}
\examples{
#library(pls)

#x <- rd_model('iris', 'Petal.Length')
#exe(x)

#plot_pred_rd(modelo.rd)

}
