% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_manipulation.R
\name{plot_RMSE}
\alias{plot_RMSE}
\title{plot_RMSE}
\usage{
plot_RMSE(model)
}
\arguments{
\item{model}{a dimension reduction model.}
}
\description{
graph the root mean square error of cross validation according to components used.
}
\examples{
#library(pls)

#x <- rd_model('iris', 'Petal.Length')
#exe(x)

#plot_RMSE(modelo.rd)

}
