% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_manipulation.R
\name{plot_var_pred_rd}
\alias{plot_var_pred_rd}
\title{plot_var_pred_rd}
\usage{
plot_var_pred_rd(
  model,
  n.comp,
  titles = c("Varianza Explicada en Variable a Predecir", "Número de Componente",
    "Porcentaje de Varianza Explicada")
)
}
\arguments{
\item{model}{a dimension reduction model.}

\item{n.comp}{the optimum number of components.}

\item{titles}{labels on the chart}
}
\value{
echarts4r plot
}
\description{
graph of the variance explained in the variable to predict according to the components used.
}
\examples{
library(pls)

x <- rd_model('iris', 'Petal.Length')
exe(x)

plot_var_pred_rd(modelo.rd,1)

}
\author{
Ariel Arroyo <luis.ariel.arroyo@promidat.com>
}
