% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tools.R
\name{radioButtonsTr}
\alias{radioButtonsTr}
\title{radioButtonsTr}
\usage{
radioButtonsTr(inputId, label, values, names)
}
\arguments{
\item{inputId}{the id of the group.}

\item{label}{label of the radio buttons group.}

\item{values}{vector with the values of each radio button.}

\item{names}{vector with the names of each radio button.}
}
\value{
shiny.tag object
}
\description{
create multples with names for language change.
}
\details{
this function only work correctly on the server side because they need the css and js file.
}
\keyword{internal}
