% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{rd_type}
\alias{rd_type}
\title{rd_type}
\usage{
rd_type(mode.rd = options_regressor("rd.mode"))
}
\arguments{
\item{mode.rd}{the method of dimension reduction is defined as mode=1 is the MCP, and mode=0 the ACP.}
}
\description{
returns the name of the method of dimension reduction.
}
\examples{
rd_type(1)
rd_type(0)

}
\seealso{
\code{\link[pls]{pcr}}, \code{\link[pls]{plsr}}
}
