% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionResult.R
\docType{class}
\name{RegressionResult-class}
\alias{RegressionResult-class}
\alias{RegressionResult}
\title{A Regression Result}
\description{
The regression result is a simple record which should include
  the output of a regression process stripped from any non-needed additional
  information.
}
\section{Slots}{

\describe{
\item{\code{time}}{the time consumed during the regression process}

\item{\code{result}}{the regression result, or \code{NULL} if regression has failed}

\item{\code{metric}}{the metric used during the regression, or \code{NULL} if the
metric is not to be stored}

\item{\code{name}}{the name of the fitted model}
}}

\seealso{
RegressionResult.new
}
