% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionResult.R
\name{RegressionResult.new}
\alias{RegressionResult.new}
\title{Helper Method to Create an Instance of \code{\link{RegressionResult}}}
\usage{
RegressionResult.new(time = 0L, result = NULL, metric = NULL)
}
\arguments{
\item{time}{the time consumed during the regression process}

\item{result}{the regression result, or \code{NULL} if regression has failed}

\item{metric}{the metric used during the regression, or \code{NULL} if the
metric is not to be stored}
}
\value{
an instance of  \code{\link{RegressionResult}}
}
\description{
Always use this function to instantiate
  \code{\link{RegressionResult}}.
}
