% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionResults.R
\docType{class}
\name{RegressionResults-class}
\alias{RegressionResults-class}
\alias{RegressionResults}
\title{A List of \code{\link{RegressionResult}}s for a Specific Setup
  Identified by a List of Names}
\description{
Each instance of this  class represents list of regression
  results for a specific setup identified by a list of names.
}
\section{Slots}{

\describe{
\item{\code{names}}{the vector of name components of this setup}

\item{\code{results}}{the list of instances of \code{\link{RegressionResult}} for
this setup}
}}

