% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionResults.R
\name{RegressionResults.new}
\alias{RegressionResults.new}
\title{Create an Instance of \code{\link{RegressionResults}}}
\usage{
RegressionResults.new(names, results)
}
\arguments{
\item{names}{the vector of name components of this setup}

\item{results}{the list of instances of \code{\link{RegressionResult}} for
this setup}
}
\description{
Instantiate \code{\link{RegressionResults}} using this method.
}
