% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchLoad.R
\name{regressoR.batchLoad}
\alias{regressoR.batchLoad}
\title{Load the \code{\link{RegressionResults}}}
\usage{
regressoR.batchLoad(source = getwd(),
  selector = path.extensionRegExp(extensions = "model", before.extension =
  "_single"), namesProcessor = identity, check.directory = NULL,
  cores = 1L, logging = if (cores <= 1L) {     TRUE } else {    
  file.path(destination, "log.txt") })
}
\arguments{
\item{source}{the source folder}

\item{selector}{the selector for the files}

\item{check.directory}{a function which can choose if a directory should be
followed or not}

\item{cores}{the number of cores to be used for loading}
}
\value{
a list of \code{\link{RegressionResults}} instances
}
\description{
Load all the setup results from a folder structure.
}
