% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPlot.R
\name{regressoR.batchPlot}
\alias{regressoR.batchPlot}
\title{Plot all the Curves from a Data Set together}
\usage{
regressoR.batchPlot(results, log = "")
}
\arguments{
\item{results}{an instance or list of \code{\link{RegressionResult}}}

\item{log}{the logarithmic scaling info for the axes}
}
\description{
A simple utility method for visualizing the data.
}
