% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultLearners.R
\name{regressoR.defaultLearners}
\alias{regressoR.defaultLearners}
\title{Get the Default Learners for Regression of 2-Dimensional Data}
\usage{
regressoR.defaultLearners()
}
\value{
a list of regression learners that can be applied.
}
\description{
Create the list of default learners for the use in
  \code{\link{regressoR.applyLearners}}. These comprise some functional
  models, some simple multi-layer perceptrons, and splines.
}
