% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnForExport.R
\name{regressoR.learnForExport}
\alias{regressoR.learnForExport}
\title{Learn a Model for Export}
\usage{
regressoR.learnForExport(x, y, learners = regressoR.defaultLearners(),
  representations = Transformation.applyDefault2D(x = x, y = y, addIdentity =
  TRUE), metricGenerator = RegressionQualityMetric.default, q = 0.75,
  includeMetric = TRUE)
}
\arguments{
\item{x}{the \code{x} coordinates, i.e., the input values}

\item{y}{the \code{y} coordinates, i.e., the output values}

\item{learners}{the learners to apply}

\item{representations}{the list of data representations, or \code{NULL} if
fitting should take place only on the raw data}

\item{metricGenerator}{the metric generator function}

\item{q}{the effort parameter: 0=minimum effort=fast/low quality, 1=maximum
effort=slow=highest quality}

\item{includeMetric}{should be the quality function to rate models be stored
in the list as well?}
}
\value{
an instance of \code{\link{RegressionResult}}
}
\description{
A version of the \code{\link{regressoR.learn}} function which
  prepares the result as instance of \code{\link{RegressionResult}} by
  stripping down all but the essential information (and potentially storing
  the original quality metric). The returned list is suitable for export into
  a file and import via batch processing.
}
\seealso{
regressoR.batchLearn
}
