% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResult.R
\name{regressoR.loadResult}
\alias{regressoR.loadResult}
\title{Load a Result Stored by \code{\link{regressoR.batchLearn}}}
\usage{
regressoR.loadResult(file)
}
\arguments{
\item{file}{the path of the file to read}
}
\value{
the loaded result
}
\description{
Open a file \code{file} produced by
  \code{\link{regressoR.batchLearn}} and return the stored result.
}
\seealso{
regressoR.batchLearn

regressoR.batchLearn
}
