% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLearners.R
\name{regressoR.makeLearners}
\alias{regressoR.makeLearners}
\title{Make all the Default Learners for Regression of 2-Dimensional Data}
\usage{
regressoR.makeLearners()
}
\value{
a list of regression learners that can be applied.
}
\description{
Create the list of default learners for the use in
  \code{\link{regressoR.applyLearners}}. Currently, these are the learners
  obtained by
  \code{\link[regressoR.functional]{FunctionalModel.makeLearners}()}.
}
