% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeResults.R
\name{regressoR.mergeResults}
\alias{regressoR.mergeResults}
\title{Create a Combination or Selection of Results}
\usage{
regressoR.mergeResults(results, nameTransformer = identity)
}
\arguments{
\item{results}{the list of instances of \code{\link{RegressionResults}}}

\item{nameTransformer}{a transformation function to be applied to the
\code{names} field of each result record}
}
\value{
a list of \code{\link{RegressionResults}} where all names have been
  transformed according to \code{nameTransformer} and where the
  \code{results} members of instances with the same name have been merged (so
  that each name only occurs once in the end)
}
\description{
Merge a set of results according to a given name transformation.
}
