% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultLearners.R
\name{regressoR.monotonousLearners}
\alias{regressoR.monotonousLearners}
\title{Get the Default Learners for Monotonous Regression of 2-Dimensional
  Data}
\usage{
regressoR.monotonousLearners()
}
\value{
a list of regression learners that can be applied.
}
\description{
Create the list of default learners for monotonous models for
  the use in \code{\link{regressoR.applyLearners}}. These comprise some
  monotonous functional models, some simple restricted multi-layer
  perceptrons, and bounded splines.
}
