% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressoR.R
\docType{package}
\name{regressoR}
\alias{regressoR-package}
\alias{regressoR}
\title{Regression Data Analysis System}
\description{
Perform a supervised data analysis on a database through a 'shiny' graphical interface. 
It includes methods such as linear regression, penalized regression, k-nearest neighbors, decision trees, ada boosting, 
extreme gradient boosting, random forest, neural networks, deep learning and support vector machines.
}
\details{
\tabular{ll}{
Package: \tab regressoR\cr
Type: \tab Package\cr
Version: \tab 4.0.2\cr
Date: \tab 2024-11-15\cr
License: \tab GPL (>=2)\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://promidat.website/}
  \item Report bugs at \url{https://github.com/PROMiDAT/predictoR/issues}
}

}
\author{
Oldemar Rodriguez Rojas \cr
Maintainer: Oldemar Rodriguez Rojas <oldemar.rodriguez@ucr.ac.cr>
}
\keyword{package}
