% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tools.R
\name{render_index_table}
\alias{render_index_table}
\title{render_index_table}
\usage{
render_index_table(table)
}
\arguments{
\item{table}{the data.frame to be converted}
}
\description{
creates a reactive table for indices panels.
}
\examples{
if(interactive()) {
   library(shiny)
   shinyApp(
      ui = fluidPage(fluidRow(column(12, tableOutput('tbl')))),
      server = function(input, output) {
         output$tbl = render_index_table(iris)
      }
   )
}

}
