% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{rf_model}
\alias{rf_model}
\title{rf_model}
\usage{
rf_model(
  data = "datos.aprendizaje",
  variable.pred = NULL,
  model.var = "modelo.rf",
  ntree = 500,
  mtry = 1
)
}
\arguments{
\item{data}{the name of the learning data.}

\item{variable.pred}{the name of the variable to be predicted.}

\item{model.var}{the name of the variable that stores the resulting model.}

\item{ntree}{the ntree parameter of the model.}

\item{mtry}{the mtry parameter of the model.}
}
\description{
generates the code to create the random forest model.
}
\examples{
library(randomForest)
x <- rf_model('iris', 'Petal.Length')
exe(x)
print(modelo.rf)

}
\seealso{
\code{\link[randomForest]{randomForest}}
}
