% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{rf_prediction}
\alias{rf_prediction}
\title{rf_prediction}
\usage{
rf_prediction(data = "datos.prueba", variable.pred = NULL,
  model.var = "modelo.rf", pred.var = "prediccion.rf")
}
\arguments{
\item{data}{the name of the test data.}

\item{variable.pred}{the name of the variable to be predicted.}

\item{model.var}{the name of the variable that stores the resulting model.}

\item{pred.var}{the name of the variable that stores the resulting prediction.}
}
\description{
generates the code to create the prediction of the random forest model.
}
\examples{
library(randomForest)
library(dplyr)

x <- rf_model('iris', 'Petal.Length', model.var = 'model_rf')
exe(x)
print(model_rf)

x <- rf_prediction('iris', 'Petal.Length', 'model_rf', 'my_prediction')
exe(x)
print(my_prediction)

}
