% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{rl_coeff}
\alias{rl_coeff}
\title{rl_coeff}
\usage{
rl_coeff(model.var = "modelo.rl")
}
\arguments{
\item{model.var}{the name of the variable that stores the resulting model.}
}
\description{
generates the code to get the information of the coefficients of the linear regression model
}
\examples{
x <- rl_model('iris', 'Petal.Length', 'model_rl')
exe(x)
print(model_rl)

x <- rl_coeff('model_rl')
exe(x)

print(df.rl)
print(r2)

}
