% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_linear_regression_utils.R
\name{rl_model}
\alias{rl_model}
\title{rl_model}
\usage{
rl_model(data, variable.pred)
}
\arguments{
\item{data}{dataframe}

\item{variable.pred}{the name of the variable to be predicted.}
}
\description{
generates a linear regression model.
}
\seealso{
\code{\link[stats]{lm}}
}
