% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_linear_regression_utils.R
\name{rl_prediction}
\alias{rl_prediction}
\title{rl_prediction}
\usage{
rl_prediction(modelo, test.data)
}
\arguments{
\item{data}{the name of the test data.}

\item{model.var}{the name of the variable that stores the resulting model.}

\item{pred.var}{the name of the variable that stores the resulting prediction.}
}
\description{
generates the code to create the prediction of the linear regression model.
}
\seealso{
\code{\link[stats]{predict}}
}
