% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{rlr_type}
\alias{rlr_type}
\title{rlr_type}
\usage{
rlr_type(alpha_rlr = options_regressor("rlr.alpha"))
}
\arguments{
\item{alpha_rlr}{the penalty is defined as alpha=1 is the lasso penalty, and alpha=0 the ridge penalty.}
}
\description{
returns the name of the penalty according to the alpha.
}
\examples{
rlr_type(1)
rlr_type(0)

}
\seealso{
\code{\link[glmnet]{glmnet}}
}
