% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{summary_indices}
\alias{summary_indices}
\title{summary_indices}
\usage{
summary_indices(data)
}
\arguments{
\item{data}{a numeric vector.}
}
\value{
a list with the summary.
}
\description{
resume una variable retornando el mínimo, el primer cuartil, el tercer cuartil y el valor máximo.
}
\examples{
summary_indices(iris$Sepal.Length)

}
