% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SVM_utils.R
\name{svm_model}
\alias{svm_model}
\title{svm_model}
\usage{
svm_model(data, variable.pred, scale = TRUE, kernel = "linear")
}
\arguments{
\item{data}{dataframe}

\item{variable.pred}{the name of the variable to be predicted.}

\item{scale}{the scale parameter of the model.}

\item{kernel}{string. The kernel parameter of the model.}
}
\description{
generates a support vector machines model.
}
\seealso{
\code{\link[e1071]{svm}}
}
