% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SVM_utils.R
\name{svm_prediction}
\alias{svm_prediction}
\title{svm_prediction}
\usage{
svm_prediction(model, test.data)
}
\arguments{
\item{model}{a support vector machine model(svm).}

\item{test.data}{dataframe.}
}
\description{
generates the prediction of the support vector machine model.
}
