% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tools.R
\name{tabsOptions}
\alias{tabsOptions}
\title{tabsOptions}
\usage{
tabsOptions(buttons = list(icon("gear"), icon("terminal")),
  widths = c(50, 100), heights = c(100, 50), tabs.content = list("",
  ""))
}
\arguments{
\item{buttons}{vector or list with icons of each tab.}

\item{widths}{vector or list with widths of each tab.}

\item{heights}{vector or list with heights of each tab.}

\item{tabs.content}{list with the content of each tab.}
}
\value{
shiny.tag
}
\description{
create tabs options on panels.
}
\details{
this function only work correctly on the server side because they need the css and js file.
}
\keyword{internal}
