% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{validate_pn_data}
\alias{validate_pn_data}
\title{validate_pn_data}
\usage{
validate_pn_data(x, y, var.pred = "", language = "es")
}
\arguments{
\item{x}{a data.frame with criteria to compare.}

\item{y}{a data.frame to be comprared.}

\item{var.pred}{a vector with the names of variables to be excluded from the comparison.}

\item{language}{the language to choose. It can be "es" or "en".}
}
\description{
Verify that a data.frame has the same columns with the same types.
}
\examples{
\dontrun{
validate_pn_data(iris, cars)
validate_pn_data(iris, iris)
x <- iris
x$Species <- as.numeric(x$Species)
validate_pn_data(iris, x)
}
}
