% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{var.categorical}
\alias{var.categorical}
\title{var.categorical}
\usage{
var.categorical(data)
}
\arguments{
\item{data}{the dataset}
}
\value{
a vector with the names of the categorical columns
}
\description{
gets only the categorical columns
}
\examples{
var.categorical(iris)

}
