% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{var.numerical}
\alias{var.numerical}
\title{var.numerical}
\usage{
var.numerical(data)
}
\arguments{
\item{data}{the dataset}
}
\value{
a vector with the names of the numerical columns
}
\description{
gets only the numerical columns
}
\examples{
var.numerical(iris)

}
