% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{var_categorical}
\alias{var_categorical}
\title{var_categorical}
\usage{
var_categorical(data)
}
\arguments{
\item{data}{the dataset.}
}
\value{
a vector with the names of the categorical columns.
}
\description{
gets only the categorical columns.
}
\examples{
var_categorical(iris)

}
