% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPlot.R
\name{batchPlot.RegressionResults}
\alias{batchPlot.RegressionResults}
\title{Plot all the Curves from a Data Set together}
\usage{
batchPlot.RegressionResults(results, ...)
}
\arguments{
\item{results}{an instance or list of \code{\link{RegressionResult}}}

\item{...}{Arguments passed on to \code{plotteR::batchPlot.list}
\describe{
  \item{plotXY}{should the \code{x-y} points be plotted (if \code{yfun} is not
\code{NULL})}
  \item{widthXY}{the line width for points to be plotted (only considered if
\code{plotXY} is \code{TRUE} and \code{yfun} is not \code{NULL})}
  \item{plotXF}{should the \code{x-y} lines be plotted (if \code{ffun} is not
\code{NULL})}
  \item{widthXF}{the line width for lines to be plotted (only considered if
\code{plotXF} is \code{TRUE} and \code{ffun} is not \code{NULL})}
  \item{names}{the names of the lines to be printed in the legend, or
\code{NULL} if no legend should be plotted}
  \item{colors}{the colors to be used for the plot}
  \item{xlab}{the label for the x-axis}
  \item{ylab}{the label for the y-axis}
  \item{legend}{a list of additional parameters to be passed to
\code{\link[graphics]{legend}}, or \code{NULL} to use the default
parameters}
  \item{x.min.lower}{a lower bound for the automatically computed \code{x}
coordinate minimum}
  \item{x.min.upper}{an upper bound for the automatically computed \code{x}
coordinate minimum}
  \item{x.max.lower}{a lower bound for the automatically computed \code{x}
coordinate maximum}
  \item{x.max.upper}{an upper bound for the automatically computed \code{x}
coordinate maximum}
  \item{y.min.lower}{a lower bound for the automatically computed \code{y}
coordinate minimum}
  \item{y.min.upper}{an upper bound for the automatically computed \code{y}
coordinate minimum}
  \item{y.max.lower}{a lower bound for the automatically computed \code{y}
coordinate maximum}
  \item{y.max.upper}{an upper bound for the automatically computed \code{y}
coordinate maximum}
  \item{x.add}{some additional \code{x} coordinates at which the function
should be evaluated, or \code{TRUE} if the \code{x} coordinate minimum and
maximum over all data sets should be added as evaluation points}
}}
}
\description{
A simple utility method for visualizing the data.
}
