% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPlot.R
\name{batchPlot.RegressionResults}
\alias{batchPlot.RegressionResults}
\title{Plot all the Curves from a Data Set together}
\usage{
batchPlot.RegressionResults(results, log = "", plotPoints = TRUE,
  plotFun = TRUE)
}
\arguments{
\item{results}{an instance or list of \code{\link{RegressionResult}}}

\item{log}{the logarithmic scaling info for the axes}

\item{plotPoints}{should the points be plotted?}

\item{plotFun}{should the fitted function models be plotted?}
}
\description{
A simple utility method for visualizing the data.
}
