% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boosting_utils.R
\name{boosting_importance_plot}
\alias{boosting_importance_plot}
\title{boosting_importance_plot}
\usage{
boosting_importance_plot(
  model,
  titles = c("Importancia de Variables segun Influencia Relativa",
    "Influencia Relativa", "Variable")
)
}
\arguments{
\item{model}{boosting model(gbm).}

\item{titles}{Labels on the chart}
}
\description{
generates the graph of variable importance.
}
