% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_boosting_utils.R
\name{boosting_model}
\alias{boosting_model}
\title{boosting_model}
\usage{
boosting_model(
  data,
  variable.pred,
  n.trees = 50,
  distribution = "gaussian",
  shrinkage = 0.1
)
}
\arguments{
\item{data}{dataframe}

\item{variable.pred}{the name of the variable to be predicted.}

\item{n.trees}{integer specifying the total number of trees to fit.}

\item{distribution}{either a character string specifying the name of the distribution to use or a list with a component name specifying the distribution and any additional parameters needed.}

\item{shrinkage}{the shrinkage parameter of the model. The learning rate or step-size reduction}
}
\description{
generates a boosting model.
}
\seealso{
\code{\link[gbm]{gbm}}
}
