% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{boosting_prediction}
\alias{boosting_prediction}
\title{boosting_prediction}
\usage{
boosting_prediction(
  data = "datos.prueba",
  variable.pred = NULL,
  model.var = "modelo.boosting",
  pred.var = "prediccion.boosting",
  n.trees = 50
)
}
\arguments{
\item{data}{the name of the test data.}

\item{variable.pred}{the name of the variable to be predicted.}

\item{model.var}{the name of the variable that stores the resulting model.}

\item{pred.var}{the name of the variable that stores the resulting prediction.}

\item{n.trees}{the n.trees parameter of the model.}
}
\description{
generates the code to create the prediction of the boosting model.
}
\examples{
library(gbm)
library(dplyr)
x <- boosting_model('iris', 'Petal.Length', "model_boosting")
exe(x)
print(model_boosting)

x <- boosting_prediction('iris', 'Petal.Length', 'model_boosting', 'my_prediction')
exe(x)
print(my_prediction)

}
\seealso{
\code{\link[gbm]{gbm}}
}
