% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tools.R
\name{code_field}
\alias{code_field}
\title{code_field}
\usage{
code_field(runid, fieldid, ...)
}
\arguments{
\item{runid}{the id of the execute button.}

\item{fieldid}{the id of the code field.}

\item{...}{arguments to be passed to \code{\link[shinyAce]{aceEditor}}}
}
\value{
shiny.tag object
}
\description{
create an \code{\link[shinyAce]{aceEditor}} wrapper with a button to execute the code
}
\details{
this function only work correctly on the server side because they need the css and js file.
}
\keyword{internal}
