% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{coef_lambda}
\alias{coef_lambda}
\title{coef_lambda}
\usage{
coef_lambda(
  data = "datos.aprendizaje",
  variable.pred = NULL,
  model.var = "modelo.rlr",
  lambda = NULL,
  cv.var = "cv.glm"
)
}
\arguments{
\item{data}{the name of the learning data.}

\item{variable.pred}{the name of the variable to be predicted.}

\item{model.var}{the name of the variable that stores the resulting model.}

\item{lambda}{a numerical value in case you don't want to use the optimal lambda.}

\item{cv.var}{the variable that stores the optimal lambda.}
}
\description{
generates the code to print the penalized regression coefficients.
}
\examples{
library(glmnet)
x <- rlr_model('iris', 'Petal.Length')
exe(x)

x <- coef_lambda('iris','Petal.Length', 'modelo.rlr')
exe(x)

}
