% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{def_code_cat}
\alias{def_code_cat}
\title{def_code_cat}
\usage{
def_code_cat(data = "datos", variable)
}
\arguments{
\item{data}{the name of the current data.}

\item{variable}{the name of the variable for the categorical distribution chart.}
}
\description{
def_code_cat
}
\examples{
x <- def_code_cat('iris', 'Species')
exe(x)

}
