% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{dt_model}
\alias{dt_model}
\title{dt_model}
\usage{
dt_model(
  data = "datos.aprendizaje",
  variable.pred = NULL,
  model.var = "modelo.dt",
  minsplit = 20,
  maxdepth = 15
)
}
\arguments{
\item{data}{the name of the learning data.}

\item{variable.pred}{the name of the variable to be predicted.}

\item{model.var}{the name of the variable that stores the resulting model.}

\item{minsplit}{the minsplit parameter of the model.}

\item{maxdepth}{the maxdepth parameter of the model.}
}
\description{
generates the code to create the decision trees model.
}
\examples{
library(rpart)

x <- dt_model('iris', 'Petal.Length')
exe(x)
print(modelo.dt)

}
\seealso{
\code{\link[rpart]{rpart}}
}
