% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{dt_prediction}
\alias{dt_prediction}
\title{dt_prediction}
\usage{
dt_prediction(
  data = "datos.prueba",
  model.var = "modelo.dt",
  pred.var = "prediccion.dt"
)
}
\arguments{
\item{data}{the name of the test data.}

\item{model.var}{the name of the variable that stores the resulting model.}

\item{pred.var}{the name of the variable that stores the resulting prediction.}
}
\description{
generates the code to create the prediction of the decision trees model.
}
\examples{
library(rpart)

x <- dt_model('iris', 'Petal.Length', model.var = 'model_dt')
exe(x)
print(model_dt)

x <- dt_prediction('iris', 'model_dt', 'my_prediction')
exe(x)
print(my_prediction)

}
