% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{exe}
\alias{exe}
\title{exe}
\usage{
exe(..., envir = parent.frame())
}
\arguments{
\item{...}{one or more texts to be concatenated and executed.}

\item{envir}{the environment in which expr is to be evaluated.}
}
\value{
the result of the execute.
}
\description{
concat and execute a text in R.
}
\examples{
exe("5+5")
exe("5","+","5")
exe("plot(iris$Species)")

}
