% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{extract_code}
\alias{extract_code}
\title{extract_code}
\usage{
extract_code(funcion, envir = parent.frame())
}
\arguments{
\item{funcion}{the name of the function to be extracted.}

\item{envir}{the environment in which expr is to be evaluated.}
}
\description{
gets the code of a function in text form.
}
\examples{
extract_code("cat")
extract_code("plot")

parse(text = extract_code("plot"))

}
