% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_r_forest_utils.R
\name{importance_plot_rf}
\alias{importance_plot_rf}
\title{importance_plot_rf}
\usage{
importance_plot_rf(
  model.rf,
  titles = c("Importancia de Variables Segun el Porcentaje de Incremento del MSE",
    "Aumento porcentual del error cuadratico medio", "Variable")
)
}
\arguments{
\item{model.rf}{a random forest model.}

\item{titles}{labels on the chart}
}
\value{
echarts4r plot
}
\description{
graphs the importance of variables for the random forest model according to the percentage increase in mean square error.
}
\seealso{
\code{\link[randomForest]{randomForest}}
}
\author{
Ariel Arroyo <luis.ariel.arroyo@promidat.com>
}
