% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_manipulation.R
\name{importance_plot_rf}
\alias{importance_plot_rf}
\title{importance_plot_rf}
\usage{
importance_plot_rf(model.rf, title.1, title.2)
}
\arguments{
\item{model.rf}{a random forest model.}

\item{title.1}{the title of the first chart.}

\item{title.2}{the title of the second chart.}
}
\description{
graphs the importance of variables for the random forest model.
}
\examples{
library(randomForest)
x <- rf_model('iris', 'Petal.Length')
exe(x)
importance_plot_rf(modelo.rf, translate('impVarA'), translate('impVarRSS'))

}
\seealso{
\code{\link[randomForest]{randomForest}}
}
