% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_regressor.R
\name{init_regressor}
\alias{init_regressor}
\title{This function will start regressoR}
\usage{
init_regressor()
}
\value{
Nothing
}
\description{
An interactive 'Shiny' application for data regression.
}
\details{
Start regressoR

This starts the regressoR application on the user's local computer.
}
\examples{
 if(interactive()){
   init_regressor()
 }
}
\keyword{regressoR}
