% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_KNN_utils.R
\name{kkn_model}
\alias{kkn_model}
\title{kkn_model}
\usage{
kkn_model(
  data,
  variable.pred,
  scale = TRUE,
  k = 7,
  kernel = "optimal",
  distance = 2
)
}
\arguments{
\item{data}{dataframe}

\item{variable.pred}{the name of the variable to be predicted.}

\item{scale}{the scale parameter of the model.}

\item{k}{the k value of the model.}

\item{kernel}{string. The kernel parameter of the model.}

\item{distance}{the distance parameter of the model.}
}
\description{
generates a k nearest neighbors model.
}
\seealso{
\code{\link[kknn]{train.kknn}}
}
