% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_KNN_utils.R
\name{kkn_prediction}
\alias{kkn_prediction}
\title{kkn_prediction}
\usage{
kkn_prediction(model, test.data)
}
\arguments{
\item{model}{k nearest neighbors model(train.kknn).}

\item{test.data}{dataframe.}
}
\description{
generates the prediction of the k nearest neighbors model.
}
