% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{kkn_prediction}
\alias{kkn_prediction}
\title{kkn_prediction}
\usage{
kkn_prediction(data = "datos.prueba", variable.pred = NULL,
  model.var = "modelo.knn", pred.var = "prediccion.knn")
}
\arguments{
\item{data}{the name of the test data.}

\item{variable.pred}{the name of the variable to be predicted.}

\item{model.var}{the name of the variable that stores the resulting model.}

\item{pred.var}{the name of the variable that stores the resulting prediction.}
}
\description{
generates the code to create the prediction of the k nearest neighbors model.
}
\examples{
library(kknn)
library(dplyr)

x <- kkn_model('iris', 'Petal.Length', model.var = 'model_knn')
exe(x)
print(model_knn)

x <- kkn_prediction('iris', 'Petal.Length', 'model_knn', 'my_prediction')
exe(x)
print(my_prediction)

}
