% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{models_mode}
\alias{models_mode}
\title{models_mode}
\usage{
models_mode(list.names = list())
}
\arguments{
\item{list.names}{a list whose names function as keys for \code{\link[regressoR]{translate}}. The names have to have the key-mode form.}
}
\description{
transforms the names of a list from key-mode form to value-mode form.
}
\examples{
x <- list('knnl-mode1' = 1, 'knnl-mode2' = 2, 'knnl-mode2' = 5)
models_mode(x)

}
