% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{models_mode}
\alias{models_mode}
\title{models_mode}
\usage{
models_mode(list.names = list(), language = "es")
}
\arguments{
\item{list.names}{a list whose names function as keys for \code{\link{translate}}. The names have to have the key-mode form.}

\item{language}{the language to choose. It can be "es" or "en".}
}
\value{
a vector with the names
}
\description{
models_mode
}
\examples{
x <- list('knnl-mode1' = 1, 'knnl-mode2' = 2, 'knnl-mode2' = 5)
models_mode(x)

}
