% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_neural_networks_utils.R
\name{nn_prediction}
\alias{nn_prediction}
\title{nn_prediction}
\usage{
nn_prediction(model, test.data)
}
\arguments{
\item{model}{neural network model(neuralnet).}

\item{test.data}{dataframe.}
}
\description{
generates the prediction of a neural network model.
}
\seealso{
\code{\link[neuralnet]{compute}}
}
