% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Predictive_Power_utils.R
\name{pairs_power}
\alias{pairs_power}
\title{pairs_power}
\usage{
pairs_power(data, decimals = 2)
}
\arguments{
\item{data}{A DataFrame}

\item{decimals}{Number of numbers after the decimal point.}
}
\description{
Generate a pair chart
}
