% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{pairs_power}
\alias{pairs_power}
\title{pairs_power}
\usage{
pairs_power(data = "datos")
}
\arguments{
\item{data}{the name of the current data.}
}
\description{
pairs_power
}
\examples{
\dontrun{
library(psych)
x <- pairs_power('iris')
exe(x)
}

}
\seealso{
\code{\link[psych]{pairs.panels}}
}
