% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_models_utils.R
\name{plot_real_prediction}
\alias{plot_real_prediction}
\title{plot_real_prediction}
\usage{
plot_real_prediction(real, pred, titles = c("Real", "Prediccion"))
}
\arguments{
\item{real}{the real values in traning-testing.}

\item{pred}{the prediction values in traning-testing.}

\item{titles}{Labels on the chart}
}
\value{
echarts4r plot
}
\description{
scatter plot between the actual value of the variable to be predicted and the prediction of the model.
}
\author{
Ariel Arroyo <luis.ariel.arroyo@promidat.com>
}
