% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_manipulation.R
\name{plot_RMSE}
\alias{plot_RMSE}
\title{plot_RMSE}
\usage{
plot_RMSE(model, n.comp = "n.comp.rd")
}
\arguments{
\item{model}{a dimension reduction model.}

\item{n.comp}{the name of the variable that stores the optimum number of components.}
}
\description{
graph the root mean square error of cross validation according to components used.
}
\examples{
library(pls)

x <- rd_model('iris', 'Petal.Length')
exe(x)

plot_RMSE(modelo.rd)

}
