% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dimension_reduction_utils.R
\name{plot_RMSE}
\alias{plot_RMSE}
\title{plot_RMSE}
\usage{
plot_RMSE(
  model,
  n.comp,
  titles = c("RMSE Segun Numero de Componentes", "Numero de Componente", "RMSE")
)
}
\arguments{
\item{model}{a dimension reduction model.}

\item{n.comp}{the optimum number of components.}

\item{titles}{labels on the chart}
}
\value{
echarts4r plot
}
\description{
graph the root mean square error of cross validation according to components used.
}
\author{
Ariel Arroyo <luis.ariel.arroyo@promidat.com>
}
