% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{rd_model}
\alias{rd_model}
\title{rd_model}
\usage{
rd_model(data = "datos.aprendizaje", variable.pred = NULL,
  model.var = "modelo.rd", n.comp = "n.comp.rd",
  mode = options_regressor("rd.mode"), scale = TRUE)
}
\arguments{
\item{data}{the name of the learning data.}

\item{variable.pred}{the name of the variable to be predicted.}

\item{model.var}{the name of the variable that stores the resulting model.}

\item{n.comp}{the name of the variable that stores the optimum number of components.}

\item{mode}{the method of dimension reduction is defined as mode=1 is the MCP, and mode=0 the ACP.}

\item{scale}{the scale parameter of the model.}
}
\description{
generates the code to create the dimension reduction model.
}
\examples{
library(pls)

x <- rd_model('iris', 'Petal.Length')
exe(x)
print(modelo.rd)

}
\seealso{
\code{\link[pls]{pcr}}, \code{\link[pls]{plsr}}
}
