% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dimension_reduction_utils.R
\name{rd_prediction}
\alias{rd_prediction}
\title{rd_prediction}
\usage{
rd_prediction(model, test.data, ncomp = NULL)
}
\arguments{
\item{model}{dimension reduction model(pcr/plsr).}

\item{test.data}{dataframe.}

\item{ncomp}{a numerical value in case you don't want to use the optimum number of components.}
}
\description{
generates the prediction of a dimension reduction model.
}
