% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchLoad.R
\name{regressoR.batchLoad}
\alias{regressoR.batchLoad}
\title{Load the \code{\link{RegressionResults}} Generated by
  \code{\link{regressoR.batchLearn}}}
\usage{
regressoR.batchLoad(source = getwd(),
  selector = path.extensionRegExp(extensions = "model", before.extension =
  "_single"), namesProcessor = identity, check.directory = NULL,
  cores = 1L, logging = if (cores <= 1L) {     TRUE } else {    
  file.path(destination, "log.txt") })
}
\arguments{
\item{source}{the source folder}

\item{selector}{the selector for the files}

\item{namesProcessor}{a function which receives a vector of name components
(extracted from the directory path of the results) and maps them to the
vector of name components to actually use (by default, identity mapping)}

\item{check.directory}{a function which can choose if a directory should be
followed or not}

\item{cores}{the number of cores to be used for loading}

\item{logging}{should progress information be printed: either \code{TRUE} for
printing to the console via \code{\link{print}}, \code{FALSE} for no
logging, or a path to a file receiving logging information}
}
\value{
a list of \code{\link{RegressionResults}} instances
}
\description{
Load all the setup results built by
  \code{\link{regressoR.batchLearn}} from a folder structure.
}
