% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_random_forests_utils.R
\name{rf_model}
\alias{rf_model}
\title{rf_model}
\usage{
rf_model(data, variable.pred, ntree = 500, mtry = 1)
}
\arguments{
\item{data}{dataframe}

\item{variable.pred}{the name of the variable to be predicted.}

\item{ntree}{the ntree parameter of the model.}

\item{mtry}{the mtry parameter of the model.}
}
\description{
generates a random forest model.
}
\seealso{
\code{\link[randomForest]{randomForest}}
}
