% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_random_forests_utils.R
\name{rf_prediction}
\alias{rf_prediction}
\title{rf_prediction}
\usage{
rf_prediction(model, test.data)
}
\arguments{
\item{model}{Random Forest model(randomForest).}

\item{test.data}{dataframe.}
}
\description{
generates the prediction of the random forest model.
}
