% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_linear_regression_utils.R
\name{rl_coeff}
\alias{rl_coeff}
\title{rl_coeff}
\usage{
rl_coeff(modelo)
}
\arguments{
\item{modelo}{linear regression model}
}
\description{
get the information of the coefficients of the linear regression model
}
