% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{rl_model}
\alias{rl_model}
\title{rl_model}
\usage{
rl_model(
  data = "datos.aprendizaje",
  variable.pred = NULL,
  model.var = "modelo.rl"
)
}
\arguments{
\item{data}{the name of the learning data.}

\item{variable.pred}{the name of the variable to be predicted.}

\item{model.var}{the name of the variable that stores the resulting model.}
}
\description{
generates the code to create the linear regression model.
}
\examples{
x <- rl_model('iris', 'Petal.Length')
exe(x)
print(modelo.rl)

}
\seealso{
\code{\link[stats]{lm}}
}
