% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{rl_prediction}
\alias{rl_prediction}
\title{rl_prediction}
\usage{
rl_prediction(
  data = "datos.prueba",
  model.var = "modelo.rl",
  pred.var = "prediccion.rl"
)
}
\arguments{
\item{data}{the name of the test data.}

\item{model.var}{the name of the variable that stores the resulting model.}

\item{pred.var}{the name of the variable that stores the resulting prediction.}
}
\description{
generates the code to create the prediction of the linear regression model.
}
\examples{
x <- rl_model('iris', 'Petal.Length', 'model_rl')
exe(x)
print(model_rl)

x <- rl_prediction('iris', 'model_rl', 'my_prediction')
exe(x)
print(my_prediction)

}
\seealso{
\code{\link[stats]{predict}}
}
