% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_linear_regression_utils.R
\name{rl_prediction}
\alias{rl_prediction}
\title{rl_prediction}
\usage{
rl_prediction(model, test.data)
}
\arguments{
\item{model}{a linear regression model(lm).}

\item{test.data}{dataframe.}
}
\description{
generates the prediction of the linear regression model.
}
\seealso{
\code{\link[stats]{predict}}
}
