% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{rlr_model}
\alias{rlr_model}
\title{rlr_model}
\usage{
rlr_model(
  data = "datos.aprendizaje",
  variable.pred = NULL,
  model.var = "modelo.rlr",
  cv.var = "cv.glm",
  alpha = 0,
  standardize = TRUE
)
}
\arguments{
\item{data}{the name of the learning data.}

\item{variable.pred}{the name of the variable to be predicted.}

\item{model.var}{the name of the variable that stores the resulting model.}

\item{cv.var}{the variable that stores the optimal lambda.}

\item{alpha}{the alpha parameter of the model.}

\item{standardize}{the standardize parameter of the model.}
}
\description{
generates the code to create the penalized regression model.
}
\examples{
library(glmnet)
x <- rlr_model('iris', 'Petal.Length')
exe(x)
print(modelo.rlr)

}
\seealso{
\code{\link[glmnet]{glmnet}}, \code{\link[glmnet]{cv.glmnet}}
}
