% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{rlr_prediction}
\alias{rlr_prediction}
\title{rlr_prediction}
\usage{
rlr_prediction(data.a = "datos.aprendizaje", data.p = "datos.prueba",
  variable.pred = NULL, model.var = "modelo.rlr",
  pred.var = "prediccion.rlr", lambda = NULL, cv.var = "cv.glm")
}
\arguments{
\item{data.a}{the name of the learning data.}

\item{data.p}{the name of the test data.}

\item{variable.pred}{the name of the variable to be predicted.}

\item{model.var}{the name of the variable that stores the resulting model.}

\item{pred.var}{the name of the variable that stores the resulting prediction.}

\item{lambda}{a numerical value in case you don't want to use the optimal lambda.}

\item{cv.var}{the variable that stores the optimal lambda.}
}
\description{
generates the code to create the prediction of the penalized regression model.
}
\examples{
library(glmnet)
x <- rlr_model('iris', 'Petal.Length')
exe(x)
print(modelo.rlr)

x <- rlr_prediction('iris', 'iris', 'Petal.Length', pred.var = 'my_prediction')
exe(x)
print(my_prediction)

}
