% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_utils.R
\name{summary_indices}
\alias{summary_indices}
\title{summary_indices}
\usage{
summary_indices(data)
}
\arguments{
\item{data}{a numeric vector.}
}
\description{
summarizes a variable by returning the minimum, first quartile, third quartile and maximum value.
}
\examples{
summary_indices(iris$Sepal.Length)

}
