% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{svm_model}
\alias{svm_model}
\title{svm_model}
\usage{
svm_model(
  data = "datos.aprendizaje",
  variable.pred = NULL,
  model.var = "modelo.svm",
  scale = TRUE,
  kernel = "linear"
)
}
\arguments{
\item{data}{the name of the learning data.}

\item{variable.pred}{the name of the variable to be predicted.}

\item{model.var}{the name of the variable that stores the resulting model.}

\item{scale}{the scale parameter of the model.}

\item{kernel}{the kernel parameter of the model.}
}
\description{
generates the code to create the support vector machines model.
}
\examples{
library(e1071)
x <- svm_model('iris', 'Petal.Length')
exe(x)
print(modelo.svm)

}
\seealso{
\code{\link[e1071]{svm}}
}
