% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generate.R
\name{svm_prediction}
\alias{svm_prediction}
\title{svm_prediction}
\usage{
svm_prediction(data = "datos.prueba", variable.pred = NULL,
  model.var = "modelo.svm", pred.var = "prediccion.svm")
}
\arguments{
\item{data}{the name of the test data.}

\item{variable.pred}{the name of the variable to be predicted.}

\item{model.var}{the name of the variable that stores the resulting model.}

\item{pred.var}{the name of the variable that stores the resulting prediction.}
}
\description{
generates the code to create the prediction of the support vector machines model.
}
\examples{
library(e1071)
library(dplyr)

x <- svm_model('iris', 'Petal.Length', model.var = 'model_svm')
exe(x)
print(model_svm)

x <- svm_prediction('iris', 'Petal.Length', 'model_svm', 'my_prediction')
exe(x)
print(my_prediction)

}
