% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{translate}
\alias{translate}
\title{translate}
\usage{
translate(text, language = "es")
}
\arguments{
\item{text}{the id for the text.}

\item{language}{the language to choose. It can be "es" or "en".}
}
\description{
translates text id into current language.
}
\examples{
translate("knn")
translate("knn", "en")

}
