% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_manipulation.R
\name{word_report}
\alias{word_report}
\title{word_report}
\usage{
word_report(title = "Sin Titulo", name = "PROMiDAT",
  order_by_regressor = TRUE, extra = "")
}
\arguments{
\item{title}{report title.}

\item{name}{name of the author of the report.}

\item{order_by_regressor}{it's the order for the default "regressoR" report.}

\item{extra}{a string with any extra code you want to add to the configuration chunk.}
}
\description{
creates a header for the report that allows you to generate a word file.
}
\examples{
new_report(iris, 'iris')

new_section_report()
insert_report('1_part', 'Title 1', 'head(iris)\\n', 'summary(iris)')

new_section_report()
insert_report('1_part', 'Title 1', 'head(iris)\\n', 'summary(iris)')

word_report(order_by_regressor = FALSE)

}
