% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_manipulation.R
\name{word_report}
\alias{word_report}
\title{word_report}
\usage{
word_report(title = "Sin Titulo", name = "PROMiDAT",
  order_by_regressor = TRUE)
}
\arguments{
\item{title}{report title.}

\item{name}{name of the author of the report.}

\item{order_by_regressor}{It's the order for the default "regressoR" report.}
}
\description{
word_report
}
